+++ src/__mocks__/data/dappConfig.ts
@@ -1,7 +1,7 @@
 export const dappConfig = {
-  id: 'devnet',
-  name: 'Devnet',
-  egldLabel: 'xEGLD',
+  id: 'sovereign',
+  name: 'Sovereign',
+  egldLabel: 'WEGLD',
   decimals: '4',
   egldDenomination: '18',
   gasPerDataByte: '1500',
@@ -9,8 +9,9 @@ export const dappConfig = {
   walletConnectDeepLink:
     'https://maiar.page.link/?apn=com.multiversx.maiar.wallet&isi=1519405832&ibi=com.multiversx.maiar.wallet&link=https://maiar.com/',
   walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
-  walletAddress: 'https://devnet-wallet.multiversx.com',
-  apiAddress: 'https://devnet-api.multiversx.com',
-  explorerAddress: 'http://devnet-explorer.multiversx.com',
-  chainId: 'D'
+  walletAddress: 'https://wallet-sovereign-test.elrond.ro/',
+  apiAddress: 'https://api-sovereign-test.elrond.ro',
+  explorerAddress:
+    'https://api-sovereign-test.internal-explorer.multiversx.com',
+  chainId: 'S'
 };