+++ src/__mocks__/data/testKeystoreWallet/index.ts
@@ -1,3 +1,5 @@
 export * from './keystoreAccount';
+export * from './keystoreWalletCollections';
 export * from './keystoreWalletNfts';
 export * from './keystoreWalletTokens';
+export * from './pendingTransactionKeystoreWallet';