+++ src/lib/index.ts
@@ -3,4 +3,3 @@ export * from './sdkDapp';
 export * from './sdkDappCore';
 export * from './sdkDappForm';
 export * from './sdkJsWebWalletIo';
-export * from './sdkWallet';