+++ src/pages/RegisterToken/hooks/useRegisterTokenOptions.ts
@@ -1,6 +1,6 @@
 import { useEffect, useMemo } from 'react';
 import { useGetTokensWithEgld } from 'hooks';
-import { useGetAccountInfo } from 'lib';
+import { useGetAccountInfo, getEgldLabel } from 'lib';
 import { useLazyGetCollectionsQuery } from 'redux/endpoints';
 import { SendTypeEnum, TokenOptionType } from 'types';
 
@@ -11,24 +11,28 @@ export const useRegisterTokenOptions = (sendType: SendTypeEnum) => {
     fetchCollections,
     { data: collections, isLoading: isLoadingCollections }
   ] = useLazyGetCollectionsQuery();
+  const egldLabel = getEgldLabel();
 
   const getTokenOptionsByType = (type: SendTypeEnum): TokenOptionType[] => {
+    let options: TokenOptionType[] = [];
+
     if (type === SendTypeEnum.nft) {
-      return (
+      options =
         collections?.map((token) => ({
           value: token.ticker,
           label: token.name
-        })) ?? []
-      );
+        })) ?? [];
+    } else {
+      options = tokens
+        .filter((token) => token.identifier !== egldLabel)
+        .map((token) => ({
+          value: token.identifier,
+          label: token.name
+        }));
     }
 
-    // Remove EGLD
-    tokens.shift();
-
-    return tokens.map((token) => ({
-      value: token.identifier,
-      label: token.name
-    }));
+    // Show only tokens/collections with a prefix (e.g. sov-FNG-123456)
+    return options.filter((token) => token.value.split('-').length > 2);
   };
 
   const getTokens = (type: SendTypeEnum) =>
@@ -36,17 +40,16 @@ export const useRegisterTokenOptions = (sendType: SendTypeEnum) => {
 
   useEffect(() => {
     fetchCollections(address);
-  }, [address, websocketEvent]);
+  }, [address, sendType, websocketEvent]);
 
   const tokenOptions = useMemo(
     () => getTokenOptionsByType(sendType),
     [collections, tokens, sendType]
   );
 
-  const allTokens = [...tokens, ...(collections || [])];
-
-  // Remove EGLD
-  allTokens.shift();
+  const allTokens = [...tokens, ...(collections || [])].filter(
+    (token) => !('identifier' in token) || token.identifier !== egldLabel
+  );
 
   return {
     allTokens,